+++ src/main/java/megabrain/gyeongnamgyeongmae/auctionItem/dto/AuctionItemResponse.java
@@ -21,7 +21,7 @@ public class AuctionItemResponse {
   private String name;
 
   private String nickname;
-  private int price;
+  private Long price;
   private AuctionItemStatus itemStatus;
   private AuctionStatus auctionStatus;
 
@@ -35,8 +35,8 @@ public class AuctionItemResponse {
 
   private String content;
 
-  private int likeCount;
-  private int viewCount;
+  private Long likeCount;
+  private Long viewCount;
 
   public static AuctionItemResponse of(AuctionItem auctionItem) {
     return AuctionItemResponse.builder()
